requirements.txt
reflex
requests
python-dotenv
pydantic
